:sparkles: Feat: 탑 네비게이션에 검색바 삽입
