+++ src/components/common/Navigation/ExplorePageNavigation.tsx
@@ -1,4 +1,12 @@
-import { BackButton } from "./BackButton";
+import { useRouter } from "next/router";
+import tw from "twin.macro";
+
+import { useScrollDirection } from "@/application/hooks";
+import { Logo } from "@/components/common/Navigation/Logo";
+import { SSRSuspense } from "@/components/common/Suspense";
+import { SearchInput } from "@/components/search";
+import { TagCategory } from "@/components/tags";
+
 import { Navigation } from "./Navigation";
 import { SideBar } from "./SideBar";
 
@@ -6,15 +14,35 @@ interface Props {
   title?: string;
 }
 export const ExplorePageNavigation = ({ title }: Props) => {
+  const router = useRouter();
+  const direction = useScrollDirection();
   return (
-    <Navigation>
-      <Navigation.Left>
-        <BackButton />
-        <div className="max-w-200 truncate font-suit text-18-bold-140">{title}</div>
-      </Navigation.Left>
-      <Navigation.Right>
-        <SideBar />
-      </Navigation.Right>
-    </Navigation>
+    <>
+      <Navigation>
+        <Navigation.Left>
+          <Logo />
+        </Navigation.Left>
+        <Navigation.Right>
+          <SideBar />
+        </Navigation.Right>
+      </Navigation>
+      <section
+        className={`sticky z-10 flex items-center gap-7 bg-white transition-[top] ${
+          direction === "DOWN" ? "top-54" : "top-0"
+        }`}
+      >
+        <SearchInput
+          css={tw`placeholder:text-gray-900`}
+          inputMode="none"
+          placeholder={title}
+          onClick={() => {
+            router.push("/search");
+          }}
+        />
+        <SSRSuspense>
+          <TagCategory />
+        </SSRSuspense>
+      </section>
+    </>
   );
 };
