+++ src/components/common/Navigation/index.ts
@@ -1,6 +1,7 @@
 export * from "./BackButtonNavigation";
 export * from "./ExplorePageNavigation";
 export * from "./IntroPageNavigation";
+export * from "./MemeDetailPageNavigation";
 export * from "./MyPageNavigation";
 export * from "./Navigation";
 export * from "./SearchPageNavigation";
